requirements.txt
streamlit~=1.36.0
python-dotenv~=1.0.0
langchain~=0.2.6
langchain-openai~=0.1.14
openai~=1.35.9
langchain-community~=0.2.6
langchainhub~=0.1.20
